+++ src/components/index.ts
@@ -1,28 +1,9 @@
-export { default as Button } from './Button'
-export { default as Input } from './Input'
-export { default as Textarea } from './Textarea'
-export { default as Text } from './Text'
-export { default as Checkbox } from './Checkbox'
-export { default as Radio } from './Radio'
-export { default as Select } from './Select'
-export { default as Window } from './Window'
-export { default as Tabs } from './Tabs'
-export { default as Modal } from './Modal'
-export { default as WindowMenu } from './WindowMenu'
-export * from './menuTypes'
-export * from './menuState'
-export * from './menuKeyboard'
-export * from './menuFocus'
-export { default as Splitter } from './Splitter'
-export { default as Collapse } from './Collapse'
-export { default as Popover } from './Popover'
-export { default as DropDownMenu } from './DropDownMenu'
-export { default as Spin } from './Spin'
-export { default as Breadcrumb } from './Breadcrumb'
-export { default as Shortcut } from './Shortcut'
-export { default as Tree } from './Tree'
-export { default as Pagination } from './Pagination'
-export { default as Transfer } from './Transfer'
-export { default as Taskbar } from './Taskbar'
-export { default as StartButton } from './StartButton'
-export { ContextMenu } from './ContextMenu'
+export const A = 1;
+
+export * from './Widget/Widget';
+export * from './Window/Window';
+export * from './Window/WindowTitle';
+export * from './Window/WindowManager';
+export * from './Screen/Screen';
+export * from './Screen/ScreenManager';
+export * from './Screen/Grid';